+++ lib/src/typedefs/tab_close_interceptor.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 
+import 'package:flutter/widgets.dart';
+
 import '../tab_data.dart';
 
 /// Intercepts a close event to indicates whether the tab can be closed.
 typedef TabCloseInterceptor = FutureOr<bool> Function(
-    int tabIndex, TabData tabData);
+    BuildContext context, int tabIndex, TabData tabData);